programFiles/requirements.txt

bs4==0.0.1
cryptography==2.1.4
gspread==3.0.0
pprint

numpy==1.13.3
oauth2client==4.1.2
pandas==0.21.1
pickleshare==0.7.4

python-dotenv


selenium==3.11.0

gitPython

PyCryptodome

beautifulsoup4==4.6.0



glob2==0.5
google-api-python-client
google-auth
